drivers: mhb: Increase diag mode size
Increase the mode parameter of the diag request command
to 32-bits to allow for UniPro testing.

Reviewed-on: http://gerrit.mot.com/842991
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted

+++ nuttx/include/nuttx/mhb/mhb_protocol.h
@@ -445,7 +445,7 @@ struct mhb_i2s_control_req {
 
 /* Diag */
 struct mhb_diag_mode_req {
-	uint8_t mode;
+	uint32_t mode;
 } __attribute__((packed));
 
 #endif
